pyproject.toml
[tool.poetry]
name = "prodigal-automation"
version = "0.1.1"
description = "Agentic tools & OAuth/RBAC support for Prodigal Automation"
authors = ["Your Name <you@company.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "prodigal_automation", from = "src" }]

[tool.poetry.dependencies]
python        = "^3.12"
pydantic      = "^2.1.1"
fastapi       = "^0.100.0"
requests      = "^2.31.0"
requests-oauthlib = "^1.3.1"
python-dotenv = "^1.1"
tweepy = "^4.15.0"
typer = "^0.16.0"
# If you want async:
# httpx = "^0.24.0"
facebook-sdk = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
black = "^24.0"
isort = "^6.0"
flake8 = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sma = "prodigal_automation.cli:app"

[tool.poetry.urls]
"Homepage"       = "https://github.com/Prodigal-AI/prodigal-automation"
"Repository"     = "https://github.com/Prodigal-AI/prodigal-automation"
"Documentation"  = "https://prodigal-automation.readthedocs.io"  # if you have docs
"Bug Tracker"    = "https://github.com/Prodigal-AI/prodigal-automation/issues"